super basic server created
