+++ package.json
@@ -5,7 +5,8 @@
   "main": "server.js",
   "dependencies": {
     "@google-cloud/vision": "^0.11.0",
-    "express": "^4.15.2"
+    "express": "^4.15.2",
+    "morgan": "^1.8.1"
   },
   "devDependencies": {},
   "scripts": {
