+++ public/index.html
@@ -1,0 +1,7 @@
+<html>
+<head>
+</head>
+<body>
+Test
+</body>
+</html>
